delete the unnecessary code and make the folder
